django_form/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.1.6
djangorestframework==3.14.0
filelock==3.9.0
idna==3.4
Pillow==9.5.0
platformdirs==3.0.0
pytz==2023.3.post1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
virtualenv==20.19.0
